python/e2e_tests/requirements.txt
# E2E Test Requirements

# Testing framework
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# HTTP client
requests>=2.31.0
httpx>=0.27.0

# Database drivers
psycopg2-binary>=2.9.9

# LLM SDKs
openai>=1.0.0
anthropic>=0.28.0
litellm>=1.30.0
groq>=0.5.0

# Frameworks
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0

# FutureAGI SDK (local install)
# pip install -e ../fi_instrumentation
# pip install -e ../frameworks/openai
# pip install -e ../frameworks/anthropic
# pip install -e ../frameworks/langchain
# pip install -e ../frameworks/litellm
# pip install -e ../frameworks/groq

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
